# api/openapi.yaml
openapi: 3.1.1
info:
  title: RelayShelf API
  version: 0.3.0
  description: HTTP contract authority for RelayShelf.
servers:
  - url: /api/v1
paths:
...
  /tags:
    get:
      operationId: listTags
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: User tags, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Tag' } } } } } }
    post:
      operationId: createTag
      security: [{ cookieAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TagRequest' } } } }
      responses: { '201': { description: Created tag, content: { application/json: { schema: { $ref: '#/components/schemas/Tag' } } } }, '409': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/Error' } }
  /tags/{tagId}:
    parameters: [{ $ref: '#/components/parameters/TagId' }]
    patch:
      operationId: updateTag
      security: [{ cookieAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateTagRequest' } } } }
      responses: { '200': { description: Updated tag, content: { application/json: { schema: { $ref: '#/components/schemas/Tag' } } } }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/Error' } }
    delete:
      operationId: deleteTag
      security: [{ cookieAuth: [] }]
      responses: { '204': { description: Deleted tag }, '404': { $ref: '#/components/responses/Error' } }
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: __Host-session }
  parameters:
...
    TagId:
      name: tagId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    TagRequest:
      type: object
      additionalProperties: false
      required: [name, color]
      properties: { name: { type: string, minLength: 1, maxLength: 64 }, color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' } }
    UpdateTagRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties: { name: { type: string, minLength: 1, maxLength: 64 }, color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' } }
    Tag:
      type: object
      required: [id, name, color, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [code, message, traceId, details]
      properties:
        code: { type: string }
        message: { type: string }
        traceId: { type: string }
        details: {}